自動アンフォロー機能追加

+++ app/Console/Commands/AutoTweetCommand.php
@@ -62,38 +62,40 @@ class AutoTweetCommand extends Command
                     Log::info('実行user_twitter_account_id:' . $userTwitterAccountId);
                     $autoTweetDatas = $t_repository->getAutoTweetDatas($userTwitterAccountId);
 
-                    foreach ($autoTweetDatas as $autoTweetData) {
-                        $highTime = $autoTweetData->tweetTime + 60;
-                        $lowTime = $autoTweetData->tweetTime - 60;
-                        Log::info('登録unixTime:' . $autoTweetData->tweetTime);
+                    if (!empty($autoTweetDatas)) {
+                        foreach ($autoTweetDatas as $autoTweetData) {
+                            $highTime = $autoTweetData->tweetTime + 60;
+                            $lowTime = $autoTweetData->tweetTime - 60;
+                            Log::info('登録unixTime:' . $autoTweetData->tweetTime);
 
-                        if (($lowTime <= $currentUnixTime) && ($currentUnixTime <= $highTime)) {
-                            Log::info('登録unixTimeは現在のunixTimeから見て前後60秒以内です');
-                            $account = $u_repository->getAccount($userTwitterAccountId);
+                            if (($lowTime <= $currentUnixTime) && ($currentUnixTime <= $highTime)) {
+                                Log::info('登録unixTimeは現在のunixTimeから見て前後60秒以内です');
+                                $account = $u_repository->getAccount($userTwitterAccountId);
 
-                            Log::info('user_id:' . $account->user_id);
-                            Log::info('screen_name:' . $account->screen_name);
+                                Log::info('user_id:' . $account->user_id);
+                                Log::info('screen_name:' . $account->screen_name);
 
-                            $response = Twitter::getAuthConnection($account->user_id, $account->screen_name)->post("statuses/update", array(
-                                "status" => $autoTweetData->tweetText,
-                            ));
+                                $response = Twitter::getAuthConnection($account->user_id, $account->screen_name)->post("statuses/update", array(
+                                    "status" => $autoTweetData->tweetText,
+                                ));
 
-                            if (isset($response->error) && $response->error != '') {
-                                return $response->error;
-                            } else {
-                                Log::info('ツイート成功');
-                                $u_repository->tweetCountSave($userTwitterAccountId);
-                                Log::info('ツイートカウントアップ');
-                                Log::info('autoTweetDataのID:' . $autoTweetData->id . 'をDBから削除します');
-                                $t_repository->deleteAutoTweetData($autoTweetData->id);
-                                Log::info('=============================');
-                                Log::info('自動ツイートアクション: メール通知');
-                                Log::info('=============================');
-                                $user = $u_repository->cronFindUser($account->user_id, $account->screen_name);
-                                Mail::to($user->email)->send(new AutoFollowMail($user));
+                                if (isset($response->error) && $response->error != '') {
+                                    return $response->error;
+                                } else {
+                                    Log::info('ツイート成功');
+                                    $u_repository->tweetCountSave($userTwitterAccountId);
+                                    Log::info('ツイートカウントアップ');
+                                    Log::info('autoTweetDataのID:' . $autoTweetData->id . 'をDBから削除します');
+                                    $t_repository->deleteAutoTweetData($autoTweetData->id);
+                                    Log::info('=============================');
+                                    Log::info('自動ツイートアクション: メール通知');
+                                    Log::info('=============================');
+                                    $user = $u_repository->cronFindUser($account->user_id, $account->screen_name);
+                                    Mail::to($user->email)->send(new AutoFollowMail($user));
+                                }
                             }
+                            Log::info('登録unixTimeは現在のunixTimeから見て前後60秒以内ではありません');
                         }
-                        Log::info('登録unixTimeは現在のunixTimeから見て前後60秒以内ではありません');
                     }
                 } else {
                     Log::info('userTwitterAccountId:' . $userTwitterAccountId . 'は実行するtweetDataがないのでauto_tweet_flgを0にします');
